move encrypter to pseudo module
