+++ logprep/util/pseudo/commands/encrypter.py
